melhorias em funcionalidades do sistema para entrega em 07-04

+++ NightPassSpr/src/main/java/br/com/fiap/NightPassSpr/Dao/EstabelecimentoDAO.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 import br.com.fiap.NightPassSpr.Entity.Estabelecimento;
 
-public interface EstabelecimentoDAO extends GenericDAO<Estabelecimento, Integer> {
+public interface EstabelecimentoDAO extends GenericDAO<Estabelecimento, Long> {
 	public List<Estabelecimento> listarPorNome();
 }
