+++ NightPassSpr/src/main/java/br/com/fiap/NightPassSpr/Dao/GeneroMusicalDAO.java
@@ -2,4 +2,4 @@ package br.com.fiap.NightPassSpr.Dao;
 
 import br.com.fiap.NightPassSpr.Entity.GeneroMusical;
 
-public interface GeneroMusicalDAO extends GenericDAO<GeneroMusical, Integer> {}
+public interface GeneroMusicalDAO extends GenericDAO<GeneroMusical, Long> {}
