+++ NightPassSpr/src/main/java/br/com/fiap/NightPassSpr/Dao/PFisicaDAO.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import br.com.fiap.NightPassSpr.Entity.PessoaFisica;
 
-public interface PFisicaDAO extends GenericDAO<PessoaFisica, Integer> {
+public interface PFisicaDAO extends GenericDAO<PessoaFisica, Long> {
 	public List<PessoaFisica> listarPorNome();
 	public PessoaFisica buscarPorEmail(String emailUsuario);
 	public boolean autenticarUsuario(String emailUsuario, String Senha);
