+++ NightPassSpr/src/main/java/br/com/fiap/NightPassSpr/Dao/PJuridicaDAO.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import br.com.fiap.NightPassSpr.Entity.PJuridica;
 
-public interface PJuridicaDAO extends GenericDAO<PJuridica, Integer> {
+public interface PJuridicaDAO extends GenericDAO<PJuridica, Long> {
   public List<PJuridica> listarPorNome();
   public PJuridica cadastrarR (PJuridica pjuridica);
 }
