+++ NightPassSpr/src/main/java/br/com/fiap/NightPassSpr/Dao/TipoEstabelecimentoDAO.java
@@ -2,4 +2,4 @@ package br.com.fiap.NightPassSpr.Dao;
 
 import br.com.fiap.NightPassSpr.Entity.TipoEstabelecimento;
 
-public interface TipoEstabelecimentoDAO extends GenericDAO<TipoEstabelecimento, Integer> {}
+public interface TipoEstabelecimentoDAO extends GenericDAO<TipoEstabelecimento, Long> {}
