+++ NightPassSpr/src/main/java/br/com/fiap/NightPassSpr/Dao/TipoGaleriaDAO.java
@@ -2,4 +2,4 @@ package br.com.fiap.NightPassSpr.Dao;
 
 import br.com.fiap.NightPassSpr.Entity.TipoGaleria;
 
-public interface TipoGaleriaDAO extends GenericDAO<TipoGaleria, Integer> {}
+public interface TipoGaleriaDAO extends GenericDAO<TipoGaleria, Long> {}
