+++ NightPassSpr/src/main/java/br/com/fiap/NightPassSpr/Dao/TipoGaleriaDAOImpl.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import br.com.fiap.NightPassSpr.Entity.TipoGaleria;
 
 @Repository
-public class TipoGaleriaDAOImpl extends GenericDAOImpl<TipoGaleria, Integer> implements
+public class TipoGaleriaDAOImpl extends GenericDAOImpl<TipoGaleria, Long> implements
 		TipoGaleriaDAO{
 
 	public TipoGaleriaDAOImpl() {
